+++ quill/src/Quill.cpp
@@ -18,12 +18,12 @@ namespace quill
 /***/
 void preallocate()
 {
-  QUILL_MAYBE_UNUSED size_t const volatile x = detail::LogManagerSingleton::instance()
-                                                 .log_manager()
-                                                 .thread_context_collection()
-                                                 .local_thread_context()
-                                                 ->spsc_queue()
-                                                 .capacity();
+  QUILL_MAYBE_UNUSED int32_t const volatile x = detail::LogManagerSingleton::instance()
+                                                  .log_manager()
+                                                  .thread_context_collection()
+                                                  .local_thread_context()
+                                                  ->spsc_queue()
+                                                  .capacity();
 }
 
 /***/
